Update designs.js
Updated grid code.  Here are instructions from mentor for the coloring code: Hi Regina, To complete the color functionality you need to do 2 things: You want to add the click eventlistener for every created cell of the grid, so we should add the event listener inside one of the for loops, as it creates the user input number cells for every column.  Then we want to use the user chosen color from the colors options, which is found in this html element: <input type="color" id="colorPicker">   So we want to link this element from the html to all grid cells, which we named its variable as 'column' and then assign the value of the input to that cell.

So we need to change "color.value" and instead use the id of the color input html element's value.
Profile image
Ibrahim E
about 10 hours ago